refactor: extract attribution position timeseries fetch
